requirements.txt
altair==4.2.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo;python_version<"3.9"
blinker==1.6.1
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.3
dbfread==2.0.7
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Fiona==1.9.2
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflate64==0.3.1
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multivolumefile==0.2.3
munch==2.5.0
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.4
pyarrow==11.0.0
pybcj==1.0.1
pycryptodomex==3.17
pydeck==0.7.1
Pygments==2.14.0
Pympler==1.0.1
pyppmd==1.0.0
pyproj==3.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.0.2
pyzstd==0.15.6
requests==2.28.2
rich==13.3.3
shapely==2.0.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.21.0
streamlit-toggle-switch==1.0.2
tenacity==8.2.2
texttable==1.6.7
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
widgetsnbextension==4.0.7
zipp==3.15.0
